+++ src/main/java/io/jenkins/infra/repository_permissions_updater/helper/TeamsHelper.java
@@ -1,4 +1,4 @@
-package io.jenkins.infra.repository_permissions_updater.launcher;
+package io.jenkins.infra.repository_permissions_updater.helper;
 
 import io.jenkins.infra.repository_permissions_updater.Definition;
 import io.jenkins.infra.repository_permissions_updater.TeamDefinition;
@@ -17,15 +17,17 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.logging.Level;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-final class TeamsHelper {
+public final class TeamsHelper {
 
     private static final Yaml YAML_TEAM_DEFINITION = new Yaml(new Constructor(TeamDefinition.class, new LoaderOptions()));
     private static final String YAML_FILE_EXTENSION = "%s.yml";
     private static final Path TEAMS_DIR = Path.of("teams");
     private static final Predicate<String> DEVELOPER_START_WITH = s -> s.startsWith("@");
+    private static final Logger LOGGER = java.util.logging.Logger.getLogger(TeamsHelper.class.getName());
 
     /**
      * Loads a team file from yaml
@@ -38,7 +40,7 @@ final class TeamsHelper {
             td.setFilePath(teamFile);
             return td;
         } catch (IOException e) {
-            throw new RuntimeException("Failed to read " + teamFile, e);
+            throw new RuntimeException("Failed to read %s".formatted(teamFile), e);
         }
     }
 
@@ -55,7 +57,7 @@ final class TeamsHelper {
         if (fileName == null) return true;
         var result = YAML_FILE_EXTENSION.formatted(name).equalsIgnoreCase(fileName.toString());
         if (!result) {
-            throw new RuntimeException("Team file should be named " + YAML_FILE_EXTENSION.formatted(name) + " instead of the current " + path.getFileName());
+            throw new RuntimeException("Team file should be named %s instead of the current %s".formatted(YAML_FILE_EXTENSION.formatted(name), path.getFileName()));
         }
         return result;
     }
@@ -64,7 +66,7 @@ final class TeamsHelper {
      * Loads all teams from the teams/ folder.
      * Always returns non-null.
      */
-    static Map<String, TeamDefinition> loadTeams() throws Exception {
+    public static Map<String, TeamDefinition> loadTeams() throws Exception {
         Set<TeamDefinition> teamsResult;
         try(var teams = Files.list(TEAMS_DIR)) {
             teamsResult = teams.map(TeamsHelper::loadTeam)
@@ -78,11 +80,11 @@ final class TeamsHelper {
      * In which case, for `@some-team` it will replace it with the developers
      * listed for the team whose name equals `some-team` under the teams/ directory.
      */
-    static void expandTeams(Definition definition, Map<String, TeamDefinition> teamsByName) {
+    public static void expandTeams(Definition definition, Map<String, TeamDefinition> teamsByName) {
         try (var developers = Arrays.stream(definition.getDevelopers())) {
             var extendDevelopers = developers.filter(DEVELOPER_START_WITH).map(s -> {
                 var team = teamsByName.get(s.substring(1));
-                ArtifactoryPermissionsUpdater.LOGGER.log(Level.INFO, "[" + definition.getName() + "]: replacing " + s + " with " + String.join(",", Arrays.asList(team.getDevelopers())));
+                LOGGER.log(Level.INFO,"[{0}]: replacing {1} with {2}", new Object[]{definition.getName(), s, String.join(",", Arrays.asList(team.getDevelopers()))});
                 return (Set<String>) new HashSet<>(Arrays.asList(team.getDevelopers()));
             }).reduce((strings, strings2) -> Stream.concat(strings.stream(), strings2.stream()).collect(Collectors.toSet()));
             var result = new HashSet<>(Arrays.asList(definition.getDevelopers()));
